+++ projects/services-layers/src/lib/types/Layers.ts
@@ -29,14 +29,19 @@ export interface ILayerStyleSet extends IAnyObject {
 }
 
 export interface popup {
-  /** limit properties */
+  /** limit layer or feature properties: only those properties of a layer/feature, that are listed in this array, are being passed through to a popup-render-function */
   filterkeys?: Array<string>;
-  /** To overwrite the keys of the layer properties */
+  /** To overwrite the keys (and only the keys) of the layer/feature properties. Object has the form {"oldKey": "newKey"} */
   properties?: IAnyObject;
   /** function to create html string - popupobj: nativeLayer */
   pupupFunktion?: (popupobj: IAnyObject) => string;
   /** async function where you can paste a html string to the callback - popupobj: nativeLayer */
   asyncPupup?: (popupobj: any, cb: (html: any) => void) => void;
+  /** create popup using angular component */
+  dynamicPopup?: {
+    component: Type<any>;
+    getAttributes?: (args: any) => object;
+  };
   /** default event is click - use move for a popup on hover */
   event?: 'move' | 'click';
   /** default is false - removes the other popups if the next is added */